Extract candidate entry and entry selection
Move the candidate list into ConfigurationCacheCandidateEntries and the
fingerprint check into ConfigurationCacheEntrySelector, leaving
DefaultConfigurationCache to coordinate them. No behaviour change.
