--- platforms/core-configuration/configuration-cache/src/main/kotlin/org/gradle/internal/cc/impl/extensions/ListExtensions.kt
@@ -1,34 +1,0 @@
-/*
- * Copyright 2024 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.gradle.internal.cc.impl.extensions
-
-
-internal
-fun <T> List<T>.withMostRecentEntry(mostRecent: T, maxEntries: Int): List<T> = when {
-    isEmpty() -> listOf(mostRecent)
-    first() == mostRecent -> this
-    else -> buildList(maxEntries) {
-        add(mostRecent)
-        val remaining = maxEntries - 1
-        if (remaining > 0) {
-            addAll(
-                this@withMostRecentEntry.asSequence()
-                    .filter { it != mostRecent }
-                    .take(remaining))
-        }
-    }
-}
